+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] (Work toward 1.1.0)
+## [Unreleased] (Work toward 1.2.0)
 ### Added
 - static folder with a `README.md` about its use
 
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Known Issues 
+
+## [1.1.0]
+### Added
+- static folder with a `README.md` about its use
+
+### Fixed
+- React Router API 
 
 ## [1.0.0] - 2017-01-04